MQLv2: emit extract() via Mqlv2IrEmitters
